Initial styles added.
