+++ src/app.scss
@@ -1,0 +1,21 @@
+@use "./styles/" as *;
+
+.app {
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: $background-color;
+}
+
+.calculator {
+  margin: 1.6rem;
+  padding: 2.4rem 1.6rem;
+  border-radius: 20px;
+  box-shadow: 0px 5px 15px 5px rgba(0, 0, 0, 0.2);
+  background-color: $primary-color-light;
+
+  @include media-tablet {
+    padding: 2.4rem 1.6rem;
+  }
+}
