+++ src/components/buttons/button.scss
@@ -1,0 +1,37 @@
+@use "../../styles/" as *;
+
+.button {
+  @include button-text;
+  width: 5rem;
+  height: 5rem;
+  border: none;
+  border-radius: 40%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  box-shadow: 2px 2px 4px 0px rgba(0, 0, 0, 0.2);
+
+  &__container {
+    padding: 0.6rem;
+  }
+
+  &--action {
+    color: $primary-color-brand;
+  }
+
+  &--operator {
+    color: $primary-color-brand;
+  }
+
+  &:hover {
+    font-weight: 600;
+    transform: translate(-1px, -1px);
+    box-shadow: 3px 3px 4px 0px rgba(0, 0, 0, 0.2);
+  }
+
+  &:active {
+    transform: translate(1px, 1px);
+    box-shadow: 1px 1px 4px 0px rgba(0, 0, 0, 0.2);
+  }
+}
