+++ src/styles/_colors.scss
@@ -1,0 +1,9 @@
+@use "sass:color";
+
+// Primary colors.
+$primary-color-dark: #13182c;
+$primary-color-light: #ffffff;
+$primary-color-brand: #68a51c;
+
+// Background colors.
+$background-color: #fafafa;
