+++ src/styles/_index.scss
@@ -1,0 +1,5 @@
+// This file enables us to @use variables and mixins within partials as a one line import.
+
+@forward "./colors";
+@forward "./typeface";
+@forward "./breakpoints";
